msg.ts -> Messages.ts file rename
