+++ src/game/gameplay/Planet.ts
@@ -3,7 +3,7 @@ import GameObject from './GameObject';
 import { play } from '../types';
 import { FrameInterval, SyncInterval, defCoords, gameField, planetRotationSpeed, planetYearAngle } from '../config';
 import { GameRoom } from '../core/Room';
-import { actionList, Classes } from '../types/msg';
+import { actionList, Classes } from '../types/Messages';
 import ObjectListManager from '../core/ListManager';
 
 export default class Planet extends GameObject {
