+++ src/game/gameplay/Ship.ts
@@ -15,7 +15,7 @@ import {
 import { GameRoom } from '../core/Room';
 import ObjectListManager from '../core/ListManager';
 import { WriteLog } from '../../database/log';
-import { PackTitle, Classes } from '../types/msg';
+import { PackTitle, Classes } from '../types/Messages';
 import { coords, rect } from '../types/gameplay';
 import Star from './Star';
 import { MoveFunction } from '../types/interfaces';
