+++ src/game/store/player.ts
@@ -1,7 +1,7 @@
 import { play } from '../types';
 import ObjectListManager from '../core/ListManager';
 import { GameRoom } from '../core/Room';
-import { actionList } from '../types/msg';
+import { actionList } from '../types/Messages';
 import GameObject from '../gameplay/GameObject';
 import { defBattleShipHealth, playerExpPerLevel, playerMaxLevel, playerStartGold } from '../config';
 import Star from '../gameplay/Star';
